X00QD: Update Wifi PowerSaveOffload
When the parameter is set to this value and voice packets are
detected in uplink or downlink data traffic, the QPower inactivity
timeout is reset to the default QPower-disabled value (200 ms).
This means that QPower power saving is not be triggered until
after 200ms of inactivity (EAPOL and ARP packets are excluded from
the check).

Change-Id: I86553a1f1dee73add4eb1b6c3933a39745a16de8